+++ src/utils/retry-policy.js
@@ -35,7 +35,7 @@ const DEFAULTS = {
  * @param {number} ms
  * @returns {Promise<void>}
  */
-export function sleep(ms) {
+function sleep(ms) {
   return new Promise((res) => setTimeout(res, ms));
 }